Bug 30120: Save and apply extended attributes during self registration verification by email
The self registration form stores a new borrower as a borrower modification when verifying by email.

Borrower modifications can  handle extended attributes.

This patch simply sotres the extended attributes in the modifications table, and approves a modification
to the extended attributes only after patron is created

To test:
1 - Apply patch
2 - Create a patron attribute and set it as viewable/editable in the OPAC
3 - Set system preference PatronSelfRegistrationVerifyByEmail
4 - Reigster a new patron on the OPAC, provide an email and populate the extended attribute
5 - Retrieve the verification token, the last on in the messages table
    SELECT * FROM message_queue;
6 - Go tot he url from above
7 - Confirm successful patron creation
8 - View patron record and confirm attribute was set

+++ opac/opac-registration-verify.pl
@@ -67,12 +67,19 @@ if (
     delete $patron_attrs->{timestamp};
     delete $patron_attrs->{verification_token};
     delete $patron_attrs->{changed_fields};
+    delete $patron_attrs->{extended_attributes};
     my $patron = Koha::Patron->new( $patron_attrs )->store;
 
     Koha::Patron::Consent->new({ borrowernumber => $patron->borrowernumber, type => 'GDPR_PROCESSING', given_on => $consent_dt })->store if $consent_dt;
 
     if ($patron) {
-        $m->delete();
+        if( $m->extended_attributes ){
+            $m->borrowernumber( $patron->borrowernumber);
+            $m->changed_fields(['extended_attributes']);
+            $m->approve();
+        } else {
+            $m->delete();
+        }
         C4::Form::MessagingPreferences::handle_form_action($cgi, { borrowernumber => $patron->borrowernumber }, $template, 1, C4::Context->preference('PatronSelfRegistrationDefaultCategory') ) if C4::Context->preference('EnhancedMessagingPreferences');
 
         $template->param( password_cleartext => $patron->plain_text_password );
